# Conflicts:
#   frontend/sswm-front/src/pages/CreateStudyRoom.jsx
#   frontend/sswm-front/src/pages/StudyRoomAdmin.jsx

+++ backend/sswm/src/main/java/com/ground/sswm/usertree/dto/UserTreeDto.java
@@ -8,7 +8,8 @@ import lombok.ToString;
 @Setter
 @ToString
 public class UserTreeDto {
+
     private Long userId;
-    private int treeId;
+    private Long treeId;
     private int exp;
 }
